+++ packages/labs/ssr/src/lib/lit-element-renderer.ts
@@ -7,7 +7,7 @@
 import {ElementRenderer} from './element-renderer.js';
 import {LitElement, CSSResult, ReactiveElement} from 'lit';
 import {_Φ} from 'lit-element/private-ssr-support.js';
-import {render, renderValue, RenderInfo} from './render-lit-html.js';
+import {render, RenderInfo} from './render-lit-html.js';
 
 export type Constructor<T> = {new (): T};
 
@@ -17,8 +17,15 @@ const {attributeToProperty, changedProperties} = _Φ;
  * ElementRenderer implementation for LitElements
  */
 export class LitElementRenderer extends ElementRenderer {
-  constructor(public element: LitElement) {
-    super(element);
+  element: LitElement;
+
+  static matchesClass(ctor: typeof HTMLElement) {
+    return ((ctor as unknown) as typeof LitElement)._$litElement$;
+  }
+
+  constructor(tagName: string) {
+    super(tagName);
+    this.element = new (customElements.get(this.tagName))();
   }
 
   connectedCallback() {
@@ -60,7 +67,7 @@ export class LitElementRenderer extends ElementRenderer {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const value = (this.element as any)?.renderLight();
     if (value) {
-      yield* renderValue(value, renderInfo);
+      yield* render(value, renderInfo);
     } else {
       yield '';
     }