+++ packages/labs/ssr/src/lib/render-with-global-dom-shim.ts
@@ -1,0 +1,8 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+import './install-global-dom-shim.js';
+export * from './render-lit-html.js';